builddeb: support creating source-only package with git archive
With git archive the bootstrap steps are not required when creating the
orig tarball, allowing the user to create a source package without
having the build-deps available at the host machine.

As debian/rules is also calling autoreconf now, it's not required to
have the cached autoconf/automake files in order to produce the build.

To create the source package simply call 'scripts/builddeb source'.

Signed-off-by: Ricardo Salveti <ricardo.salveti@linaro.org>